frontend/requirements.txt
numpy==2.2.3
Pillow==11.1.0
Requests==2.32.3
streamlit==1.42.2
